+++ pydidas_plugins/proc_plugins/pyfai_azimuthal_sector_integration.py
@@ -1,6 +1,6 @@
 # This file is part of pydidas.
 #
-# Copyright 2023, Helmholtz-Zentrum Hereon
+# Copyright 2023 - 2024, Helmholtz-Zentrum Hereon
 # SPDX-License-Identifier: GPL-3.0-only
 #
 # pydidas is free software: you can redistribute it and/or modify
@@ -21,7 +21,7 @@ several azimuthal sectors at once.
 """
 
 __author__ = "Malte Storm"
-__copyright__ = "Copyright 2023, Helmholtz-Zentrum Hereon"
+__copyright__ = "Copyright 2023 - 2024, Helmholtz-Zentrum Hereon"
 __license__ = "GPL-3.0-only"
 __maintainer__ = "Malte Storm"
 __status__ = "Production"
@@ -118,6 +118,7 @@ class PyFAIazimuthalSectorIntegration(pyFAIintegrationBase):
                 )
                 for _, _ in enumerate(self._config["sector_centers"])
             ]
+            self._ai = self._ais[0]
         self.load_and_set_mask()
         if self._mask is not None:
             for _ai in self._ais: